requirements.txt
tensorflow-gpu==1.12.0
torch==1.0.0
torchvision==0.2.1
Pillow==5.2.0
matplotlib==2.2.2
scipy==1.1.0
tqdm==4.24.0
opencv-python==3.4.3.18
pyyaml==3.13
addict==2.2.0
tensorboardX==1.4
mtcnn==0.0.8